requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Environment variables
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# AI/LLM
openai==1.35.0
anthropic==0.28.0

# Machine Learning (optional - fallback implementations provided)
scikit-learn==1.3.2
numpy==1.24.4
pandas==2.1.4
joblib==1.3.2

# System Monitoring (optional - fallback implementations provided)
psutil==5.9.6

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
httpx==0.25.2

# Telegram Bot
python-telegram-bot==20.7

# Utilities
loguru==0.7.2
rich==13.7.0

# Security and validation
cryptography==41.0.7
email-validator==2.1.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Production
gunicorn==21.2.0
redis==5.0.1
celery==5.3.4

# Note: The AI system is designed to work with or without optional dependencies
# The system will gracefully fallback to basic implementations if packages are missing
# If scikit-learn is not available, it will use mock implementations
# If psutil is not available, it will use mock system monitoring
